type fixes

+++ src/joLinearFunctionConstructors/joCurvelet2D.jl
@@ -3,7 +3,7 @@
 ## helper module
 module joCurvelet2D_etc
     using JOLI: jo_convert
-    function apply_fdct2Dwrap_real(v::AbstractVector,n1::Integer,n2::Integer,m::Integer,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
+    function apply_fdct2Dwrap_real(v::AbstractVector,n1::Integer,n2::Integer,m::Int128,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
         C=Vector{Cdouble}(m)
         eltype(v)<:Real || throw(joLinearFunctionException("joCurvelt2D: input vector must be real for real transform"))
         X=jo_convert(Cdouble,v,false)
@@ -13,7 +13,7 @@ module joCurvelet2D_etc
         C=jo_convert(rdt,C,false)
         return C
     end
-    function apply_ifdct2Dwrap_real(v::AbstractVector,n1::Integer,n2::Integer,m::Integer,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
+    function apply_ifdct2Dwrap_real(v::AbstractVector,n1::Integer,n2::Integer,m::Int128,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
         X=Vector{Cdouble}(n1*n2)
         eltype(v)<:Real || throw(joLinearFunctionException("joCurvelt2D: input vector must be real for real transform"))
         C=jo_convert(Cdouble,v,false)
@@ -23,7 +23,7 @@ module joCurvelet2D_etc
         X=jo_convert(rdt,X,false)
         return X
     end
-    function apply_fdct2Dwrap_cplx(v::AbstractVector,n1::Integer,n2::Integer,m::Integer,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
+    function apply_fdct2Dwrap_cplx(v::AbstractVector,n1::Integer,n2::Integer,m::Int128,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
         C=Vector{Complex{Cdouble}}(m)
         X=jo_convert(Complex{Cdouble},v,false)
         ccall((:jl_fdct_wrapping_cpx,:libdfdct_wrapping),Void,
@@ -32,7 +32,7 @@ module joCurvelet2D_etc
         C=jo_convert(rdt,C,false)
         return C
     end
-    function apply_ifdct2Dwrap_cplx(v::AbstractVector,n1::Integer,n2::Integer,m::Integer,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
+    function apply_ifdct2Dwrap_cplx(v::AbstractVector,n1::Integer,n2::Integer,m::Int128,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
         X=Vector{Complex{Cdouble}}(n1*n2)
         C=jo_convert(Complex{Cdouble},v,false)
         ccall((:jl_ifdct_wrapping_cpx,:libdfdct_wrapping),Void,
